requirements.txt
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
crypto==1.4.1
cryptography==3.4.7
firebase-admin==5.0.2
Flask==2.0.1
Flask-Markdown==0.3
gcloud==0.18.3
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-firestore==2.2.0
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpcio==1.39.0
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jws==0.1.3
Markdown==3.3.4
MarkupSafe==2.0.1
msgpack==1.0.2
mypy==0.910
mypy-extensions==0.4.3
Naked==0.1.31
oauth2client==4.1.3
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase4==4.5.0
python-dotenv==0.19.0
python-jwt==2.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
shellescape==3.8.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
